+++ src/components/Cards/ButtonCard/ButtonCard.stories.tsx
@@ -35,8 +35,9 @@ function ExampleDocs() {
         device, the code below is all you need to control your device.
       </p>
       <p>
-        This will automatically extract the friendly name, icon, state and group
-        of the entity to render the ButtonCard below
+        This will automatically extract the friendly name, icon, last updated, state and group
+        of the entity to render the ButtonCard below, if there's no icon linked in home assistant it will
+        use a predefined default by domain.
       </p>
       <Template />
       <h3>Source Code</h3>
@@ -59,7 +60,7 @@ function Render(args: Args) {
 }
 
 export default {
-  title: "COMPONENTS/Buttons/ButtonCard",
+  title: "COMPONENTS/Cards/ButtonCard",
   component: ButtonCard,
   tags: ["autodocs"],
   parameters: {