+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useHass } from "./useHass";
 export { useEntity } from "./useEntity";
 export { useApi } from "./useApi";
-export { useIconByDomain, useIcon } from "./useIcon";
+export { useIconByDomain, useIcon, useIconByEntity } from "./useIcon";
+export { useTimeDifference } from "./useTimeDifference";