+++ stories/HassConnectFake.tsx
@@ -32,12 +32,13 @@ interface HassProviderProps {
 const ENTITIES: HassEntities = {
   'light.fake_light': {
     attributes: {
-      friendly_name: 'Dining room light'
+      friendly_name: 'Dining room light',
+      icon: "mdi:light-recessed"
     },
     state: 'on',
     entity_id: 'light.fake_light',
-    last_changed: Date.now().toString(),
-    last_updated: Date.now().toString(),
+    last_changed: new Date().toISOString(),
+    last_updated: new Date().toISOString(),
     context: {
       id: '',
       user_id: null,
@@ -50,8 +51,8 @@ const ENTITIES: HassEntities = {
     },
     state: 'off',
     entity_id: 'switch.fake_gaming_switch',
-    last_changed: Date.now().toString(),
-    last_updated: Date.now().toString(),
+    last_changed: new Date().toISOString(),
+    last_updated: new Date().toISOString(),
     context: {
       id: '',
       user_id: null,
@@ -64,14 +65,33 @@ const ENTITIES: HassEntities = {
     },
     state: 'off',
     entity_id: 'media_player.fake_tv',
-    last_changed: Date.now().toString(),
-    last_updated: Date.now().toString(),
+    last_changed: new Date().toISOString(),
+    last_updated: new Date().toISOString(),
     context: {
       id: '',
       user_id: null,
       parent_id: null,
     }
-  }
+  },
+  'scene.good_morning': {
+    entity_id: "scene.goodmorning",
+    state: "2023-07-04T05:12:47.585217+00:00",
+    attributes: {
+        entity_id: [
+            "light.all_office_lights_2"
+        ],
+        id: "1688447567040",
+        icon: "mdi:sun-clock",
+        friendly_name: "Goodmorning"
+    },
+    context: {
+        id: "01H4FP0HY3K7R49GJS574K5TPW",
+        parent_id: null,
+        user_id: null
+    },
+    last_changed: "2023-07-04T05:29:10.851Z",
+    last_updated: "2023-07-04T05:29:10.851Z"
+}
 }
 
 function HassProvider({
@@ -98,15 +118,31 @@ function HassProvider({
   const callService = useCallback(
     async <T extends DomainName, M extends DomainService<T>>({
       service,
+      domain,
       target,
     }: CallServiceArgs<T, M>) => {
       if (typeof target !== 'string') return;
+      const dates = {
+        last_changed: new Date().toISOString(),
+        last_updated: new Date().toISOString(),
+      }
+      if (domain === 'scene') {
+        return setEntities({
+          ...entities,
+          [target]: {
+            ...entities[target],
+            ...dates,
+            state: new Date().toISOString()
+          }
+        });
+      }
       switch(service) {
         case 'turn_on':
           return setEntities({
             ...entities,
             [target]: {
               ...entities[target],
+              ...dates,
               state: 'on'
             }
           });
@@ -115,6 +151,7 @@ function HassProvider({
             ...entities,
             [target]: {
               ...entities[target],
+              ...dates,
               state: 'on'
             }
           })
@@ -123,6 +160,7 @@ function HassProvider({
             ...entities,
             [target]: {
               ...entities[target],
+              ...dates,
               state: entities[target].state === 'on' ? 'off' : 'on'
             }
           })